requirements.txt
requests~=2.32.3
uvicorn~=0.30.1
fastapi~=0.111.0
websockets~=12.0
python-dotenv~=1.0.1
alembic~=1.13.2